merge: upstream/dev #40
