+++ src/main/java/com/douzone/server/dto/vehicle/impl/VehicleEmpResDTO.java
@@ -1,6 +1,6 @@
 package com.douzone.server.dto.vehicle.impl;
 
-import com.douzone.server.dto.vehicle.IVehicleEmpResDTO;
+import com.douzone.server.dto.vehicle.jpainterface.IVehicleEmpResDTO;
 import com.douzone.server.entity.Vehicle;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
